+++ resources/lib/user_data.py
@@ -1,7 +1,6 @@
 class UserData:
-    def __init__(self, username, password, quality, site):
+    def __init__(self, username, password, site):
         self.username = username
         self.password = password
-        self.quality = quality
         self.site = 'https://www.' + site
         self.site_mobile = 'https://m.' + site